+++ src/assets/goo.png
