+++ src/types/typs.ts
@@ -1,0 +1,1 @@
+export type Hand = "rock" | "paper" | "scissors";
